+++ src/main/java/de/ukbonn/mwtek/dashboard/configuration/AcuwaveSearchConfiguration.java
@@ -78,4 +78,7 @@ public class AcuwaveSearchConfiguration extends SearchConfiguration {
 
   /** Orbis internal keys for influenza laboratory codes. */
   private List<Integer> influenzaOrbisLabPcrCodes = new ArrayList<>();
+
+  /** Reducing the procedure calls on certain wards */
+  private List<String> wardsCovidInfluenza = new ArrayList<>();
 }